Refactor: Atomic CSS width and height options
In Atomic-CSS/utility-mixins.less, simplify the name of atomic mixins for specifying widths and heights and add an em-based articulation of height specifications. Then refactor Atomic-CSS/utility-classes.less to use the names of the updated atomic mixins. Finally, in both files, simplify the format of file-sectioning comment blocks to work better with automation in code editors, such as the features of the AutoWrap package in Sublime Text 3.