Refactor: Add color override to WDS color theme
In Atomic-CSS/utility-classes.less, when utility classes for the WDS-based color theme are in use, add an override of the crimson coloring classes associated with a Spine Theme. This causes the shades of crimson associated with the old WSU branding scheme to be updated to those in the 2021 brand update.