Redesign: Presentation of gravity forms
Create a new optional ruleset for presenting gravity forms in a manner compatible with the Web Design System 2.0 theme. The new design matches aspects of the contact form present on the recently updated brand.wsu.edu website and the new UMarCom Web and Digital developer style guide. It is also significantly leaner than the previous ruleset for styling gravity forms.